Fixing bad cast (#21451)

+++ src/darwin/Framework/CHIP/MTRDeviceController.mm
@@ -577,7 +577,7 @@ static NSString * const kErrorCSRValidation = @"Extracting public key from CSR f
     }
 
     if (!chip::CanCastTo<uint32_t>(setupPIN) || !chip::SetupPayload::IsValidSetupPIN(static_cast<uint32_t>(setupPIN))) {
-        MTR_LOG_ERROR("Error: Setup pin %lu is not valid", setupPIN);
+        MTR_LOG_ERROR("Error: Setup pin %lu is not valid", static_cast<unsigned long>(setupPIN));
         if (error) {
             *error = [MTRError errorForCHIPErrorCode:CHIP_ERROR_INVALID_INTEGER_VALUE];
         }
